chore: add type check utils and separate components
In response to PR review, add proper type checking util methods, and
separate column components where multiple types are being handled.
