+++ packages/renderer/src/lib/container/ContainerColumnAge.svelte
@@ -1,11 +1,14 @@
 <script lang="ts">
 import StateChange from '../ui/StateChange.svelte';
+import { ContainerUtils } from './container-utils';
 import type { ContainerGroupInfoUI, ContainerInfoUI } from './ContainerInfoUI';
 
 export let object: ContainerInfoUI | ContainerGroupInfoUI;
+
+const containerUtils = new ContainerUtils();
 </script>
 
-{#if 'state' in object}
+{#if containerUtils.isContainerInfoUI(object)}
   <div class="text-sm text-[var(--pd-table-body-text)]">
     <StateChange state="{object.state}">{object.uptime}</StateChange>
   </div>
