+++ packages/renderer/src/lib/container/ContainerColumnImage.svelte
@@ -1,10 +1,13 @@
 <script lang="ts">
+import { ContainerUtils } from './container-utils';
 import type { ContainerGroupInfoUI, ContainerInfoUI } from './ContainerInfoUI';
 
 export let object: ContainerInfoUI | ContainerGroupInfoUI;
+
+const containerUtils = new ContainerUtils();
 </script>
 
-{#if 'image' in object}
+{#if containerUtils.isContainerInfoUI(object)}
   <div class="text-sm text-[var(--pd-table-body-text)] overflow-hidden text-ellipsis" title="{object.image}">
     {object.shortImage}
   </div>
