+++ packages/renderer/src/lib/container/ContainerColumnStatus.svelte
@@ -1,13 +1,16 @@
 <script lang="ts">
 import PodIcon from '../images/PodIcon.svelte';
 import StatusIcon from '../images/StatusIcon.svelte';
+import { ContainerUtils } from './container-utils';
 import type { ContainerGroupInfoUI, ContainerInfoUI } from './ContainerInfoUI';
 
 export let object: ContainerInfoUI | ContainerGroupInfoUI;
+
+const containerUtils = new ContainerUtils();
 </script>
 
-{#if 'type' in object}
+{#if containerUtils.isContainerGroupInfoUI(object)}
   <StatusIcon icon="{PodIcon}" status="{object.status}" />
-{:else if 'state' in object}
+{:else if containerUtils.isContainerInfoUI(object)}
   <StatusIcon icon="{object.icon}" status="{object.state}" />
 {/if}
