+++ packages/renderer/src/lib/container/container-utils.spec.ts
@@ -23,7 +23,7 @@ import type { ViewInfoUI } from '/@api/view-info';
 
 import { ContextUI } from '../context/context';
 import { ContainerUtils } from './container-utils';
-import { ContainerGroupInfoTypeUI } from './ContainerInfoUI';
+import { ContainerGroupInfoTypeUI, type ContainerGroupInfoUI, type ContainerInfoUI } from './ContainerInfoUI';
 
 let containerUtils: ContainerUtils;
 
@@ -315,4 +315,24 @@ test('should expect imageHref to use exact image id if no sha256: prefix', async
   } as unknown as ContainerInfo;
   const containerUI = containerUtils.getContainerInfoUI(containerInfo);
   expect(containerUI.imageHref).toBe('/images/dummy-sha256/dummy-engine-id/dummy-base-64/summary');
+});
+
+test('should be able to identify container groups', async () => {
+  const containerGroupInfo = {
+    id: 'my-pod',
+    name: 'My pod',
+    type: ContainerGroupInfoTypeUI.POD,
+  } as unknown as ContainerGroupInfoUI;
+  expect(containerUtils.isContainerGroupInfoUI(containerGroupInfo)).toBe(true);
+  expect(containerUtils.isContainerInfoUI(containerGroupInfo)).toBe(false);
+});
+
+test('should be able to identify containers', async () => {
+  const containerInfo = {
+    id: 'container1',
+    name: 'a container',
+    state: 'RUNNING',
+  } as unknown as ContainerInfoUI;
+  expect(containerUtils.isContainerInfoUI(containerInfo)).toBe(true);
+  expect(containerUtils.isContainerGroupInfoUI(containerInfo)).toBe(false);
 });
